hacpm/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.4
python-multipart==0.0.18
dateparser==1.2.1
recurrent==0.4.1
python-dateutil==2.9.0
Pillow==11.1.0
aiohttp==3.11.11
aiofiles==24.1.0
websockets==14.1
